requirements.txt
albumentations==0.5.2
efficientnet_pytorch==0.7.1
fastprogress==1.0.3
Flask==1.1.4
numpy==1.23.4
pandas==1.3.4
Pillow==9.2.0
scikit_learn==1.1.2
https://download.pytorch.org/whl/cpu/torch-1.12.1%2Bcpu-cp39-cp39-linux_x86_64.whl
gunicorn==20.1.0
opencv-python-headless